publish v1.8.1 - fix fractional rollout percentages rounding down (#23)
* Dont round fractional pass percentages in go

* cleanup

* bump v1.8.1

+++ evaluator.go
@@ -261,7 +261,7 @@ func evalPassPercent(user User, rule configRule, spec configSpec) bool {
 	}
 	hash := getHashUint64Encoding(spec.Salt + "." + ruleSalt + "." + getUnitID(user, rule.IDType))
 
-	return hash%10000 < (uint64(rule.PassPercentage) * 100)
+	return float64(hash%10000) < (rule.PassPercentage * 100)
 }
 
 func getUnitID(user User, idType string) string {
